This commit was generated by cvs2svn to compensate for changes in r16295, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@16296 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/monitor/mon_parse.c
@@ -2292,7 +2292,7 @@ yyreduce:
 
   case 116:
 #line 445 "mon_parse.y"
-    { (monitor_cpu_type.mon_register_set_val)(reg_memspace(yyvsp[-2].i), reg_regid(yyvsp[-2].i), yyvsp[0].i); }
+    { (monitor_cpu_type.mon_register_set_val)(reg_memspace(yyvsp[-2].i), reg_regid(yyvsp[-2].i), (WORD) yyvsp[0].i); }
     break;
 
   case 117:
